+++ tests/testthat/test-render-snapshots.R
@@ -16,7 +16,7 @@ test_that("README: marked-up grouped table", {
       Note = "Entry C3PO altered to test characters that have a special meaning in LaTeX.",
       Source = "R package \\texttt{dplyr}"
     ) |>
-    set_alignment(height:birth_year ~ "X[r]") |>
+    set_colspec(height:birth_year ~ "X[r]") |>
     set_column_labels(
       name = "", height = "Height", mass = "Mass", birth_year = "Birth Year"
     ) |>
@@ -27,8 +27,8 @@ test_that("README: marked-up grouped table", {
   expect_snapshot(cat(as_latex(x)))
 })
 
-test_that("set_alignment example", {
-  x <- tblr(countries()) |> set_alignment(value = "X", starts_with("c") ~ "c")
+test_that("set_colspec example", {
+  x <- tblr(countries()) |> set_colspec(value = "X", starts_with("c") ~ "c")
   expect_snapshot(cat(as_latex(x)))
 })